New implementations for Conflict Graphs and Bron-Kerbosch algorithm.

+++ CoinUtils/src/bron_kerbosch.h
@@ -12,10 +12,10 @@ void bk_free(BronKerbosch **_bk);
 int bk_run(BronKerbosch *bk);
 
 const CliqueSet* bk_get_clq_set(const BronKerbosch *bk);
-size_t bk_get_max_weight(const BronKerbosch *bk);
+int bk_get_max_weight(const BronKerbosch *bk);
 
-void bk_set_min_weight(BronKerbosch *bk, size_t minWeight);
-void bk_set_max_it(BronKerbosch *bk, size_t maxIt);
+void bk_set_min_weight(BronKerbosch *bk, int minWeight);
+void bk_set_max_it(BronKerbosch *bk, int maxIt);
 
 
 #endif
